Moved server-client mdc things to mdc package
