+++ reactor-aeron-benchmarks/src/main/java/reactor/aeron/AeronPongServer.java
@@ -1,5 +1,8 @@
 package reactor.aeron;
 
+import reactor.aeron.mdc.AeronResources;
+import reactor.aeron.mdc.AeronServer;
+
 public final class AeronPongServer {
 
   /**
