+++ reactor-aeron-benchmarks/src/main/java/reactor/aeron/ServerServerSends.java
@@ -6,6 +6,8 @@ import java.nio.charset.Charset;
 import java.time.Duration;
 import org.agrona.DirectBuffer;
 import org.agrona.concurrent.UnsafeBuffer;
+import reactor.aeron.mdc.AeronResources;
+import reactor.aeron.mdc.AeronServer;
 import reactor.core.publisher.Flux;
 
 public class ServerServerSends {
