+++ reactor-aeron-benchmarks/src/main/java/reactor/aeron/ServerThroughput.java
@@ -1,5 +1,8 @@
 package reactor.aeron;
 
+import reactor.aeron.mdc.AeronResources;
+import reactor.aeron.mdc.AeronServer;
+
 public class ServerThroughput {
 
   /**
