+++ reactor-aeron-benchmarks/src/main/java/reactor/aeron/rsocket/aeron/RSocketAeronClientTps.java
@@ -2,14 +2,10 @@ package reactor.aeron.rsocket.aeron;
 
 import io.rsocket.Payload;
 import io.rsocket.RSocket;
-import io.rsocket.RSocketFactory;
-import io.rsocket.frame.decoder.PayloadDecoder;
-import io.rsocket.reactor.aeron.AeronClientTransport;
 import io.rsocket.util.ByteBufPayload;
-import reactor.aeron.AeronClient;
-import reactor.aeron.AeronResources;
 import reactor.aeron.Configurations;
 import reactor.aeron.RateReporter;
+import reactor.aeron.mdc.AeronResources;
 
 public final class RSocketAeronClientTps {
 
@@ -31,19 +27,20 @@ public final class RSocketAeronClientTps {
             .start()
             .block();
 
-    RSocket client =
-        RSocketFactory.connect()
-            .frameDecoder(PayloadDecoder.ZERO_COPY)
-            .transport(
-                () ->
-                    new AeronClientTransport(
-                        AeronClient.create(resources)
-                            .options(
-                                Configurations.MDC_ADDRESS,
-                                Configurations.MDC_PORT,
-                                Configurations.MDC_CONTROL_PORT)))
-            .start()
-            .block();
+    RSocket client = null;
+    // todo io.rsocket.transport.ClientTransport was changed
+    // RSocketFactory.connect()
+    //     .frameDecoder(PayloadDecoder.ZERO_COPY)
+    //     .transport(
+    //         () ->
+    //             new AeronClientTransport(
+    //                 AeronClient.create(resources)
+    //                     .options(
+    //                         Configurations.MDC_ADDRESS,
+    //                         Configurations.MDC_PORT,
+    //                         Configurations.MDC_CONTROL_PORT)))
+    //     .start()
+    //     .block();
 
     RateReporter reporter = new RateReporter();
