+++ reactor-aeron/src/main/java/reactor/aeron/mdc/AeronClient.java
@@ -1,4 +1,4 @@
-package reactor.aeron;
+package reactor.aeron.mdc;
 
 import java.util.function.Function;
 import java.util.function.UnaryOperator;
