+++ reactor-aeron/src/main/java/reactor/aeron/mdc/AeronConnection.java
@@ -1,6 +1,9 @@
-package reactor.aeron;
+package reactor.aeron.mdc;
 
 import org.reactivestreams.Subscription;
+import reactor.aeron.AeronInbound;
+import reactor.aeron.AeronOutbound;
+import reactor.aeron.OnDisposable;
 import reactor.core.CoreSubscriber;
 import reactor.core.Disposable;
 import reactor.core.publisher.BaseSubscriber;
