+++ reactor-aeron/src/main/java/reactor/aeron/mdc/AeronExceptions.java
@@ -1,4 +1,4 @@
-package reactor.aeron;
+package reactor.aeron.mdc;
 
 class AeronExceptions {
