+++ reactor-aeron/src/main/java/reactor/aeron/mdc/DuplexAeronConnection.java
@@ -1,9 +1,12 @@
-package reactor.aeron;
+package reactor.aeron.mdc;
 
 import java.util.function.Function;
 import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import reactor.aeron.AeronInbound;
+import reactor.aeron.AeronOutbound;
+import reactor.aeron.OnDisposable;
 import reactor.core.publisher.Mono;
 import reactor.core.publisher.MonoProcessor;
