+++ reactor-aeron/src/test/java/reactor/aeron/mdc/AeronClientTest.java
@@ -1,4 +1,4 @@
-package reactor.aeron;
+package reactor.aeron.mdc;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static reactor.aeron.DefaultFragmentMapper.asString;
@@ -20,6 +20,9 @@ import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import reactor.aeron.BaseAeronTest;
+import reactor.aeron.OnDisposable;
+import reactor.aeron.SocketUtils;
 import reactor.core.publisher.BaseSubscriber;
 import reactor.core.publisher.DirectProcessor;
 import reactor.core.publisher.Flux;
