+++ reactor-aeron/src/test/java/reactor/aeron/mdc/AeronServerTest.java
@@ -1,4 +1,4 @@
-package reactor.aeron;
+package reactor.aeron.mdc;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static reactor.aeron.DefaultFragmentMapper.asString;
@@ -8,9 +8,18 @@ import java.util.function.Function;
 import java.util.stream.Stream;
 import org.agrona.DirectBuffer;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.reactivestreams.Publisher;
+import reactor.aeron.BaseAeronTest;
+import reactor.aeron.OnDisposable;
+import reactor.aeron.SocketUtils;
+import reactor.aeron.ThreadWatcher;
+import reactor.aeron.mdc.AeronClient;
+import reactor.aeron.mdc.AeronConnection;
+import reactor.aeron.mdc.AeronResources;
+import reactor.aeron.mdc.AeronServer;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.ReplayProcessor;
 import reactor.test.StepVerifier;
@@ -80,7 +89,7 @@ class AeronServerTest extends BaseAeronTest {
 
     server.dispose();
 
-    assertTrue(new ThreadWatcher().awaitTerminated(5000, "single-", "parallel-"));
+    Assertions.assertTrue(new ThreadWatcher().awaitTerminated(5000, "single-", "parallel-"));
   }
 
   private AeronConnection<DirectBuffer> createConnection() {
